update surface and price type from integer to double

+++ src/main/java/com/qrupemlak/backoffice/data/AnnouncementInfo.java
@@ -24,9 +24,9 @@ public class AnnouncementInfo {
 	private Integer floor;
 	private Boolean lift;
 	private HomeType homeType;
-	private Integer surface ;
+	private Double surface ;
 	private Integer buildingAge;
-	private Integer price;
+	private Double price;
 	private Currency currency ;
 	private Set<Image> images ;
 	private AddressInfo addressInfo ;
@@ -130,13 +130,13 @@ public class AnnouncementInfo {
 	/**
 	 * @return the surface
 	 */
-	public Integer getSurface() {
+	public Double getSurface() {
 		return surface;
 	}
 	/**
 	 * @param surface the surface to set
 	 */
-	public void setSurface(Integer surface) {
+	public void setSurface(Double surface) {
 		this.surface = surface;
 	}
 	/**
@@ -154,13 +154,13 @@ public class AnnouncementInfo {
 	/**
 	 * @return the price
 	 */
-	public Integer getPrice() {
+	public Double getPrice() {
 		return price;
 	}
 	/**
 	 * @param price the price to set
 	 */
-	public void setPrice(Integer price) {
+	public void setPrice(Double price) {
 		this.price = price;
 	}
 	/**
